[Nova] Generate Rust bindings from OpenRM or Nouveau headers
Nouveau already includes a cleaned up version of the required OpenRM
headers in drivers/gpu/drm/nouveau/include/nvrm which we can use to
generate Rust bindings for GSP data structures.

Alternatively the original OpenRM headers can be used to generate
Rust bindings from, so provide a Kconfig option to specify where these
headers are.

A Kconfig option is also provided to limit the symbols for which
bindings are generated, significantly reducing the size of the generated
Rust bindings.