rust: add new `num` module with useful integer operations
Introduce the `num` module, featuring the `NumExt` extension trait
that expands unsigned integers with useful operations for the kernel.

These are to be used by the nova-core driver, but they are so ubiquitous
that other drivers should be able to take advantage of them as well.

The currently implemented operations are:

- align_down()
- align_up()
- fls()

But this trait is expected to be expanded further.

`NumExt` is on unsigned types using a macro. An approach using another
trait constrained by the operator traits that we need (`Add`, `Sub`,
etc) was also considered, but had to be dropped as we need to use
wrapping operations, which are not provided by any trait.

Co-developed-by: Joel Fernandes <joelagnelf@nvidia.com>
Signed-off-by: Joel Fernandes <joelagnelf@nvidia.com>
Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>