gpu: nova-core: consolidate bindings and remove accessors
Wrapping the bindings in a new type adds layers and code without adding
any extra functionality. Instead just use a type alias and access the
fields directly. This results in both less code and the removal of an
unsafe statement.

If a type ever changes such that we need to hide the internal GSP
representation from the GSP code we can always just update the alias
to a new type, perhaps using a transparent representation and adding
accessors. But lets not do that for every type now - none of the types
we currently use have ever changed since OpenRm was released and adding
setters and getters for every field adds a lot of boilerplate.